+++ src/outputs/sun.h
@@ -4,14 +4,14 @@
 class Sun
 {
 private:
-    int _pin;
-    int _start;
-    int _target;
-    int _last;
+    uint8_t _pin;
+    uint8_t _start;
+    uint8_t _target;
+    uint8_t _last;
     unsigned long _start_time;
 
 public:
-    Sun(int pin);
+    Sun(uint8_t pin);
     void setup();
     void loop(const State &state);
